[migrations]: Changed migrations price and quantity to 64 bit

+++ migrations/20180118201125_rename_stockprice_to_close.up.sql
@@ -1,2 +1,2 @@
 ALTER TABLE StockHistory
-CHANGE stockPrice close int(11);
+CHANGE stockPrice close bigint(11);
